Add files via upload
Classification by computer of handwritten digits is a standard problem in pattern recognition. The typical application is automatic reading of zip codes on envelopes. In this assignment, you'll address the following problem: Given a set of manually classified digits (the training set), classify a set of unknown digits (the test set)using the SVD method.

Data SetWe will be using the US Postal Service database that contains 1707trainingand 2007 test digits (uploaded to Canvas). Each image is a grayscale 16x16 image that is converted to a 256x1 column vector by stacking all the columns of each image matrix above each other.
•The training images are stored in trainInput.csv.(256x1707).
 •The correct digit corresponding to each column of train inputs stored in trainOutput.csv.(1x1707).•The test images are stored in testInput.csv.(256x2007).
 •The correct digit corresponding to each column of test inputs stored in testOutput.csv.(1x2007).